+++ docs/conf.py
@@ -26,6 +26,7 @@ intersphinx_mapping = {
     "numpy": ("https://docs.scipy.org/doc/numpy/", None),
     "scipy": ("https://docs.scipy.org/doc/scipy/reference", None),
     "discretize": ("https://discretize.simpeg.xyz/en/master", None),
+    "empymod": ("https://empymod.readthedocs.org/en/stable", None),
 }
 
 # ==== 2. General Settings ====