+++ tests/test_utils.py
@@ -1,6 +1,7 @@
 import re
 import os
 import pytest
+import empymod
 import numpy as np
 from scipy import constants
 from timeit import default_timer
@@ -398,7 +399,7 @@ def test_TensorMesh():
 
 
 # MODEL AND FIELD CLASSES
-def test_Model():
+def test_Model(capsys):
     # Mainly regression tests
 
     # Create some dummy data
@@ -411,8 +412,11 @@ def test_Model():
     res_z = res_x*1.4
     mu_r = res_x*1.11
 
-    # Using defaults
-    model1 = utils.Model(grid)
+    _, _ = capsys.readouterr()  # Clean-up
+    # Using defaults; check backwards compatibility for freq.
+    model1 = utils.Model(grid, freq=1)
+    out, _ = capsys.readouterr()
+    assert '``Model`` does not take frequency' in out
     assert_allclose(model1.res_x, model1.res_y)
     assert_allclose(model1.nC, grid.nC)
     assert_allclose(model1.vnC, grid.vnC)
@@ -842,6 +846,160 @@ def test_grid2grid():
         utils.grid2grid(grid, field, cgrid, method='volume')
 
 
+# TIME DOMAIN
+class TestFourier:
+    def test_defaults(self, capsys):
+        time = np.logspace(-2, 2)
+        fmin = 0.01
+        fmax = 100
+
+        Fourier = utils.Fourier(time, fmin, fmax)
+        out, _ = capsys.readouterr()
+
+        assert Fourier.every_x_freq is None
+        assert Fourier.fmin == fmin
+        assert Fourier.fmax == fmax
+        assert Fourier.ft == 'ffht'
+        assert Fourier.ftarg[1] == -1.0   # Convolution DLF
+        assert Fourier.ftarg[2] == 'sin'  # Sine-DLF is default
+        assert Fourier.signal == 0        # Impulse respons
+        assert_allclose(time, Fourier.time, 0, 0)
+        assert Fourier.verb == 3          # Verbose by default
+        assert 'Key 201 CosSin (2012)' in out
+        assert 'Req. freq' in out
+        assert 'Calc. freq' in out
+        assert Fourier.freq_calc.min() >= fmin
+        assert Fourier.freq_calc.max() <= fmax
+
+        # Check frequencies to extrapolate.
+        assert_allclose(Fourier.freq_extrapolate,
+                        Fourier.freq_req[Fourier.freq_req < fmin])
+
+        # If not freq_inp nor every_x_freq, interpolate and calc have to be the
+        # same.
+        assert_allclose(Fourier.freq_interpolate, Fourier.freq_calc)
+
+        # Change time, ensure it changes required frequencies.
+        freq_req = Fourier.freq_req
+        time2 = np.logspace(-1, 2)
+        Fourier.time = time2
+        assert freq_req.size != Fourier.freq_req.size
+
+    def test_kwargs(self, capsys):
+        time = np.logspace(-1, 1)
+        fmin = 0.1
+        fmax = 10
+        freq_inp = np.logspace(-1, 1, 11)
+        xfreq = 10
+
+        # freq_inp; verb=0
+        _, _ = capsys.readouterr()
+        Fourier1 = utils.Fourier(time, fmin, fmax, freq_inp=freq_inp, verb=0)
+        out, _ = capsys.readouterr()
+        assert '' == out
+        assert_allclose(freq_inp, Fourier1.freq_calc, 0, 0)
+
+        # freq_inp AND every_x_freq => re-sets every_x_freq.
+        Fourier2 = utils.Fourier(time, fmin, fmax, every_x_freq=xfreq,
+                                 freq_inp=freq_inp, verb=1)
+        out, _ = capsys.readouterr()
+        assert 'Re-setting `every_x_freq=None`' in out
+        assert_allclose(freq_inp, Fourier2.freq_calc, 0, 0)
+        assert_allclose(Fourier1.freq_calc, Fourier2.freq_calc)
+
+        # Now set every_x_freq again => re-sets freq_inp.
+        Fourier2.every_x_freq = xfreq
+        out, _ = capsys.readouterr()
+        assert 'Re-setting `freq_inp=None`' in out
+        assert_allclose(Fourier2.freq_coarse, Fourier2.freq_req[::xfreq])
+        assert Fourier2.freq_inp is None
+        test = Fourier2.freq_req[::xfreq][
+                (Fourier2.freq_req[::xfreq] >= fmin) &
+                (Fourier2.freq_req[::xfreq] <= fmax)]
+        assert_allclose(Fourier2.freq_calc, test)
+
+        # And back
+        Fourier2.freq_inp = freq_inp
+        out, _ = capsys.readouterr()
+        assert 'Re-setting `every_x_freq=None`' in out
+        assert_allclose(Fourier2.freq_calc, freq_inp)
+        assert Fourier2.every_x_freq is None
+
+        # Unknown argument, must fail with TypeError.
+        with pytest.raises(TypeError):
+            utils.Fourier(time, fmin, fmax, does_not_exist=0)
+
+    def test_setters(self, capsys):
+        time = np.logspace(-1.4, 1.4)
+        fmin = 0.1
+        fmax = 10
+
+        # freq_inp; verb=0
+        _, _ = capsys.readouterr()
+        Fourier1 = utils.Fourier(time, fmin=np.pi/10, fmax=np.pi*10)
+        Fourier1.fmin = fmin
+        Fourier1.fmax = fmax
+        Fourier1.signal = -1
+        Fourier1.fourier_arguments('fftlog', {'pts_per_dec': 5})
+        assert Fourier1.ft == 'fftlog'
+        assert Fourier1.ftarg[0] == 5
+        assert Fourier1.ftarg[3] == -0.5  # cosine, as signal == -1
+
+    def test_interpolation(self, capsys):
+        time = np.logspace(-2, 1, 201)
+        model = {'src': [0, 0, 0], 'rec': [900, 0, 0], 'res': 1,
+                 'depth': [], 'verb': 1}
+        Fourier = utils.Fourier(time, 0.005, 10)
+
+        # Calculate data.
+        data_true = empymod.dipole(freqtime=Fourier.freq_req, **model)
+        data = empymod.dipole(freqtime=Fourier.freq_calc, **model)
+
+        # Interpolate.
+        data_int = Fourier.interpolate(data)
+
+        # Compare, extrapolate < 0.05; interpolate equal.
+        assert_allclose(data_int[Fourier.freq_extrapolate_i].imag,
+                        data_true[Fourier.freq_extrapolate_i].imag, rtol=0.05)
+        assert_allclose(data_int[Fourier.freq_calc_i].imag,
+                        data_true[Fourier.freq_calc_i].imag)
+
+        # Now set every_x_freq and again.
+        Fourier.every_x_freq = 2
+
+        data = empymod.dipole(freqtime=Fourier.freq_calc, **model)
+
+        # Interpolate.
+        data_int = Fourier.interpolate(data)
+
+        # Compare, extrapolate < 0.05; interpolate < 0.01.
+        assert_allclose(data_int[Fourier.freq_extrapolate_i].imag,
+                        data_true[Fourier.freq_extrapolate_i].imag, rtol=0.05)
+        assert_allclose(data_int[Fourier.freq_interpolate_i].imag,
+                        data_true[Fourier.freq_interpolate_i].imag, rtol=0.01)
+
+    def test_freq2transform(self, capsys):
+        time = np.linspace(0.1, 10, 101)
+        x = 900
+        model = {'src': [0, 0, 0], 'rec': [x, 0, 0], 'res': 1,
+                 'depth': [], 'verb': 1}
+
+        # Initiate Fourier instance.
+        Fourier = utils.Fourier(time, 0.001, 100)
+
+        # Calculate required frequencies.
+        data = empymod.dipole(freqtime=Fourier.freq_calc, **model)
+
+        # Transform the data.
+        tdata = Fourier.freq2time(data, x)
+
+        # Calculate data in empymod.
+        data_true = empymod.dipole(freqtime=time, signal=0, **model)
+
+        # Compare.
+        assert_allclose(data_true, tdata, rtol=1e-4)
+
+
 # FUNCTIONS RELATED TO TIMING
 def test_Time():
     t0 = default_timer()  # Create almost at the same time a
@@ -968,7 +1126,3 @@ def test_report(capsys):
         _ = utils.Report()
         out, _ = capsys.readouterr()  # Empty capsys
         assert 'WARNING :: `emg3d.Report` requires `scooby`' in out
-
-
-def test_interp3d():
-    pass